合并：
・航空支援で落ちる不具合を修正
・艦隊：改装レベル・艦載機熟練度の表示を実装(仮)
・一部地域でCSVの日付が読めなくなる不具合を修正
・戦闘：夜戦時に表示が乱れることがある不具合を修正
・情報：演習経験値予測のずれを修正

+++ ElectronicObserver/Window/FormInformation.cs
@@ -151,7 +151,7 @@ namespace ElectronicObserver.Window {
 					expbase = 500.0 + Math.Sqrt( expbase - 500.0 );
 
 				sb.AppendLine( "獲得経験値 : " + (int)expbase );
-				sb.AppendLine( "S勝利 : " + (int)( expbase * 1.2 ) );
+				sb.AppendLine( "S勝利 : " + (int)( (int)expbase * 1.2 ) );
 
 			}
